pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "baseferrhin"
version = "0.1.0"
description = "Inventaire normalisé des sites de l'âge du Fer du Rhin supérieur (Alsace, Bade-Wurtemberg, Bâle)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "pydantic>=2.0",
    "httpx>=0.27",
    "tenacity>=8.0",
    "pdfplumber>=0.11",
    "rapidfuzz>=3.0",
    "geopy>=2.4",
    "lxml>=5.0",
    "geopandas>=1.0",
    "sqlite-utils>=3.36",
    "rich>=13.0",
    "pyyaml>=6.0",
    "openpyxl>=3.1",
    "pytesseract>=0.3",
    "Pillow>=10.0",
]

[project.optional-dependencies]
ui = [
    "dash>=2.14",
    "dash-bootstrap-components>=1.5",
]
viz = [
    "keplergl>=0.3",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "ruff>=0.4",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.hatch.build.targets.wheel]
packages = ["src"]
